refactor: bind domain events to named locals; clear all compiler warnings
Two things the clang-format sweep and the resulting full rebuild exposed.

Readability
-----------
Publishing an event as a temporary with designated initialisers formats badly —
clang-format cannot keep `}.to_event())` on one line, so all nineteen call sites
came out as:

        }
                                .to_event());

Binding the event to a named local first reads better than the original and
formats cleanly. It also documents itself: `const events::UserRoleChanged
user_role_changed_event{...}` states what is being published before saying where.

Warnings
--------
The full rebuild recompiled units that had been cached for weeks, surfacing 14
warnings — 13 of them pre-existing and simply never re-emitted since. Now zero
across src/, include/ and benchmarks/ under -Wall -Wextra -Wpedantic -Wshadow
-Wconversion -Wsign-conversion.

* auth_controller: the logout-all event skipped a middle member in a designated
  initialiser, tripping -Wmissing-field-initializers. `session_id` is now set
  explicitly — "all sessions" genuinely has no single session id, and saying so
  beats leaving a reader to wonder.
* pg_*_repository (13 sites): pqxx's result::size_type and affected_rows() are
  signed; feeding them to reserve() and to a std::size_t accumulator is a
  sign-conversion. Cast explicitly.
* reaction_service::list called [[nodiscard]] authorize() purely for its
  throw-on-denied side effect. Marked (void) with a comment, since discarding
  the returned message is deliberate here and not an oversight.

Verified: full rebuild clean, 287 tests pass, 7 skipped (they need PostgreSQL),
0 failed. clang-format reports no diff.

+++ src/repositories/pg_attachment_repository.cpp
@@ -86,7 +86,7 @@ std::vector<models::Attachment> PgAttachmentRepository::list_for_message(std::in
                                 " FROM attachments WHERE message_id = $1 ORDER BY id ASC",
                             message_id);
         std::vector<models::Attachment> out;
-        out.reserve(result.size());
+        out.reserve(static_cast<std::size_t>(result.size()));
         for (const auto& row : result) {
             out.push_back(map_row(row));
         }
@@ -110,7 +110,7 @@ std::size_t PgAttachmentRepository::link_to_message(const std::vector<std::int64
                 id,
                 message_id,
                 owner_id);
-            linked += result.affected_rows();
+            linked += static_cast<std::size_t>(result.affected_rows());
         }
         return linked;
     });
